+++ tests/unit/hooks/pre-pull.test.sh
@@ -23,13 +23,13 @@ beforeAll()
 
 afterAll()
 {
-    cd - > /dev/null || exit 1
-
-    rm -rf $TEST
+    unset FLINT_HOOKS
 
     unset FLINT_CONFIG
 
-    unset FLINT_HOOKS
+    cd - > /dev/null || exit 1
+
+    [[ -n "$TEST" && -d "$TEST" ]] && rm -r $TEST
 }
 
 
@@ -106,9 +106,9 @@ unmock()
 {
     unset -f git
 
-    rm $first
+    [[ -f "$first" ]] && rm $first
 
-    rm $second
+    [[ -f "$second" ]] && rm $second
 }
 
 
@@ -262,5 +262,5 @@ it_uses_correct_git_commands()
 
     unmock
 
-    rm $commands
+    [[ -f "$commands" ]] && rm $commands
 }
