+++ tests/unit/hooks/pre-push.test.sh
@@ -15,7 +15,7 @@ afterAll()
 {
     cd - > /dev/null || exit 1
 
-    rm -rf $TEST
+    [[ -n "$TEST" && -d "$TEST" ]] && rm -r $TEST
 }
 
 
@@ -148,5 +148,5 @@ it_uses_correct_git_commands()
 
     unmock
 
-    rm $commands
+    [[ -f "$commands" ]] && rm $commands
 }
